products and providers page filters refactoring
