+++ react/assets/forus-webshop/scss/includes/components/_label.scss
@@ -1,9 +1,3 @@
-.label-group {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 5px;
-}
-
 .label {
     font: 600 11px/18px var(--base-font);
     display: flex;
