+++ react/assets/forus-webshop/scss/includes/components/label-groups.scss
@@ -1,0 +1,9 @@
+.label-group {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px 10px;
+
+    .label {
+        margin: 0 0 !important;
+    }
+}
