+++ react/assets/forus-webshop/scss/webshop.scss
@@ -25,6 +25,7 @@
 @use 'includes/components/form';
 @use 'includes/components/ui-controls';
 @use 'includes/components/label';
+@use 'includes/components/label-groups';
 @use 'includes/components/map-card';
 @use 'includes/components/tooltip';
 @use 'includes/components/triangle';
