+++ react/src/webshop/components/pages/products/Products.tsx
@@ -1,238 +1,54 @@
-import React, { Fragment, useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { useProductService } from '../../../services/ProductService';
-import { PaginationData, ResponseError } from '../../../../dashboard/props/ApiResponses';
-import Product from '../../../props/models/Product';
-import useSetProgress from '../../../../dashboard/hooks/useSetProgress';
+import React, { Fragment, useCallback, useEffect, useRef, useState } from 'react';
 import useAuthIdentity from '../../../hooks/useAuthIdentity';
 import SelectControl from '../../../../dashboard/components/elements/select-control/SelectControl';
 import CmsBlocks from '../../elements/cms-blocks/CmsBlocks';
-import useAppConfigs from '../../../hooks/useAppConfigs';
 import Paginator from '../../../../dashboard/modules/paginator/components/Paginator';
 import ProductsList from '../../elements/lists/products-list/ProductsList';
 import EmptyBlock from '../../elements/empty-block/EmptyBlock';
 import useTranslate from '../../../../dashboard/hooks/useTranslate';
 import BlockShowcaseList from '../../elements/block-showcase/BlockShowcaseList';
-import useFilterNext from '../../../../dashboard/modules/filter_next/useFilterNext';
-import { BooleanParam, NumberParam, StringParam, NumericArrayParam } from 'use-query-params';
 import { clickOnKeyEnter } from '../../../../dashboard/helpers/wcag';
-import useSetTitle from '../../../hooks/useSetTitle';
-import UIControlText from '../../../../dashboard/components/elements/forms/ui-controls/UIControlText';
-import ProductsFilterGroupReservationOptions from './elements/ProductsFilterGroupReservationOptions';
 import classNames from 'classnames';
-import ProductsFilterGroupPriceType from './elements/ProductsFilterGroupPriceType';
-import FormGroup from '../../elements/forms/FormGroup';
-import ProductsFilterGroupProductCategories from './elements/ProductsFilterGroupProductCategories';
-import ProductsFilterGroupDistance from './elements/ProductsFilterGroupDistance';
-import ProductsFilterGroupPrice from './elements/ProductsFilterGroupPrice';
-import ProductsFilterGroupFunds from './elements/ProductsFilterGroupFunds';
-import ProductsFilterGroupProviders from './elements/ProductsFilterGroupProviders';
 import { WebshopRoutes } from '../../../modules/state_router/RouterBuilder';
-import useLabelFilters from '../../../hooks/useLabelFilters';
-import useRootProductCategories from './hooks/useRootProductCategories';
-import Organization from '../../../../dashboard/props/models/Organization';
-import { useOrganizationService } from '../../../../dashboard/services/OrganizationService';
-import { useFundService } from '../../../services/FundService';
-import Fund from '../../../props/models/Fund';
-
-export type BaseTypeFilterProducts = {
-    fund_ids: number[];
-    organization_id: number;
-    product_category_ids: number[];
-    postcode: string;
-    distance: number;
-    from: number;
-    to: number;
-    qr?: boolean;
-    reservation?: boolean;
-    extra_payment?: boolean;
-    regular?: boolean;
-    discount_fixed?: boolean;
-    discount_percentage?: boolean;
-    free?: boolean;
-    informational?: boolean;
-    payout?: boolean;
-};
-
-export type TypeFilterProducts = BaseTypeFilterProducts & {
-    q: string;
-    page: number;
-    bookmarked: boolean;
-    display_type: 'list' | 'grid';
-    order_by: 'created_at' | 'price' | 'most_popular' | 'name' | 'randomized';
-    order_dir: 'asc' | 'desc';
-};
+import useProductsPageFilters from './hooks/useProductsPageFilters';
+import ProductsSidebarFilters from './elements/ProductsSidebarFilters';
+import useAppConfigs from '../../../hooks/useAppConfigs';
+import useSetTitle from '../../../hooks/useSetTitle';
+import useSetProgress from '../../../../dashboard/hooks/useSetProgress';
+import { useProductService } from '../../../services/ProductService';
+import { PaginationData, ResponseError, ResponseErrorData } from '../../../../dashboard/props/ApiResponses';
+import Product from '../../../props/models/Product';
 
 export default function Products() {
     const appConfigs = useAppConfigs();
     const authIdentity = useAuthIdentity();
-
-    const fundService = useFundService();
     const productService = useProductService();
-    const organizationService = useOrganizationService();
-
     const setTitle = useSetTitle();
-    const translate = useTranslate();
     const setProgress = useSetProgress();
-
-    const [sortByOptions] = useState(productService.getSortOptions(translate));
-
-    const [errors, setErrors] = useState<{ [key: string]: string | Array<string> }>({});
-
+    const translate = useTranslate();
+    const [errors, setErrors] = useState<ResponseErrorData>({});
     const [toMax, setToMax] = useState(0);
-
-    const [funds, setFunds] = useState<Array<Fund>>(null);
-    const [providers, setProviders] = useState<Array<Organization>>(null);
-    const { productCategoriesIconMap, productCategories } = useRootProductCategories();
-
+    const [products, setProducts] = useState<PaginationData<Product, { price_max: number }>>(null);
     const activeFetchListRef = useRef<XMLHttpRequest | null>(null);
 
-    const defaultSortOption = useMemo(() => {
-        return sortByOptions?.find((option) => {
-            return (
-                `${option.value.order_by}_${option.value.order_dir}` === appConfigs.products_default_sorting ||
-                `${option.value.order_by}` === appConfigs.products_default_sorting
-            );
-        });
-    }, [appConfigs?.products_default_sorting, sortByOptions]);
-
-    const [filterValues, filterValuesActive, filterUpdate, filter] = useFilterNext<TypeFilterProducts>(
-        {
-            q: '',
-            page: 1,
-            fund_ids: [],
-            organization_id: null,
-            product_category_ids: [],
-            postcode: '',
-            distance: null,
-            from: 0,
-            to: null,
-            qr: false,
-            reservation: false,
-            extra_payment: false,
-            bookmarked: false,
-            regular: false,
-            discount_fixed: false,
-            discount_percentage: false,
-            free: false,
-            informational: false,
-            payout: false,
-            display_type: 'grid',
-            order_by: (defaultSortOption || sortByOptions[0])?.value.order_by,
-            order_dir: (defaultSortOption || sortByOptions[0])?.value.order_dir,
-        },
-        {
-            throttledValues: ['q', 'from', 'to', 'qr', 'reservation', 'extra_payment', 'postcode'],
-            queryParams: {
-                q: StringParam,
-                page: NumberParam,
-                fund_ids: NumericArrayParam,
-                organization_id: NumberParam,
-                product_category_ids: NumericArrayParam,
-                postcode: StringParam,
-                distance: NumberParam,
-                from: NumberParam,
-                to: NumberParam,
-                qr: BooleanParam,
-                reservation: BooleanParam,
-                extra_payment: BooleanParam,
-                bookmarked: BooleanParam,
-                display_type: StringParam,
-                order_by: StringParam,
-                order_dir: StringParam,
-                regular: BooleanParam,
-                discount_fixed: BooleanParam,
-                discount_percentage: BooleanParam,
-                free: BooleanParam,
-                informational: BooleanParam,
-                payout: BooleanParam,
-            },
-            filterParams: ['display_type'],
-        },
-    );
-
-    const { labelsBlock } = useLabelFilters(filter, productCategories, funds, providers, null, toMax);
-
-    const countFiltersApplied = useMemo(() => {
-        return [
-            filterValues.q,
-            filterValues.fund_id,
-            filterValues.organization_id,
-            filterValues.product_category_id,
-        ].filter((value) => value).length;
-    }, [filterValues]);
-
-    const [products, setProducts] = useState<PaginationData<Product, { price_max: number }>>(null);
-
-    const buildQuery = useCallback(
-        (
-            values: Partial<{
-                q: string;
-                page: number;
-                fund_ids: number[];
-                organization_id: number;
-                product_category_id: number;
-                product_category_ids: number[];
-                postcode: string;
-                distance: number;
-                from: number;
-                to: number;
-                bookmarked: boolean;
-                qr: boolean;
-                reservation: boolean;
-                extra_payment: boolean;
-                regular: boolean;
-                discount_fixed: boolean;
-                discount_percentage: boolean;
-                free: boolean;
-                informational: boolean;
-                payout: boolean;
-                order_by: 'created_at' | 'price' | 'most_popular' | 'name' | 'randomized';
-                order_dir: 'asc' | 'desc';
-            }>,
-        ) => {
-            const hasFilters =
-                values.qr ||
-                values.extra_payment ||
-                values.reservation ||
-                values.regular ||
-                values.discount_fixed ||
-                values.discount_percentage ||
-                values.free ||
-                values.informational ||
-                values.payout;
-
-            return {
-                q: values.q,
-                page: values.page,
-                fund_ids: values.fund_ids?.length > 0 ? values.fund_ids : null,
-                organization_id: values.organization_id,
-                product_category_ids: values.product_category_ids?.length > 0 ? values.product_category_ids : null,
-                postcode: values.postcode || '',
-                distance: values.distance || null,
-                from: values.from || null,
-                to: values.to || null,
-                qr: hasFilters ? (values.qr ? 1 : 0) : 0,
-                reservation: hasFilters ? (values.reservation ? 1 : 0) : 0,
-                extra_payment: hasFilters ? (values.extra_payment ? 1 : 0) : 0,
-                regular: hasFilters ? (values.regular ? 1 : 0) : 0,
-                discount_fixed: hasFilters ? (values.discount_fixed ? 1 : 0) : 0,
-                discount_percentage: hasFilters ? (values.discount_percentage ? 1 : 0) : 0,
-                free: hasFilters ? (values.free ? 1 : 0) : 0,
-                informational: hasFilters ? (values.informational ? 1 : 0) : 0,
-                payout: hasFilters ? (values.payout ? 1 : 0) : 0,
-                bookmarked: values.bookmarked ? 1 : 0,
-                order_by: values.order_by,
-                order_dir: values.order_dir,
-            };
-        },
-        [],
-    );
+    const {
+        countFiltersApplied,
+        filter,
+        filterUpdate,
+        filterValues,
+        funds,
+        productCategories,
+        productCategoriesIconMap,
+        productsQuery,
+        providers,
+        sortByOptions,
+    } = useProductsPageFilters();
 
     const fetchProducts = useCallback(
         (query: object) => {
             activeFetchListRef.current?.abort();
 
-            setErrors(null);
+            setErrors({});
             setProgress(0);
 
             productService
@@ -246,46 +62,20 @@ export default function Products() {
                         return;
                     }
 
-                    setErrors(e.data?.errors);
+                    setErrors(e.data?.errors || {});
                 })
                 .finally(() => setProgress(100));
         },
         [productService, setProgress],
     );
 
-    const fetchProviders = useCallback(() => {
-        setProgress(0);
-
-        organizationService
-            .list({ type: 'provider', per_page: 300, order_by: 'name' })
-            .then((res) => setProviders(res.data.data))
-            .then(() => setProgress(100));
-    }, [organizationService, setProgress]);
-
-    const fetchFunds = useCallback(() => {
-        setProgress(0);
-
-        fundService
-            .list({ has_providers: 1 })
-            .then((res) => setFunds(res.data.data))
-            .finally(() => setProgress(100));
-    }, [fundService, setProgress]);
-
     useEffect(() => {
-        fetchFunds();
-    }, [fetchFunds]);
-
-    useEffect(() => {
-        fetchProviders();
-    }, [fetchProviders]);
-
-    useEffect(() => {
-        fetchProducts(buildQuery(filterValuesActive));
+        fetchProducts(productsQuery);
 
         return () => {
             activeFetchListRef.current?.abort();
         };
-    }, [fetchProducts, buildQuery, filterValuesActive]);
+    }, [fetchProducts, productsQuery]);
 
     useEffect(() => {
         setTitle(translate('page_state_titles.products'));
@@ -300,93 +90,18 @@ export default function Products() {
                 { name: translate('products.breadcrumbs.products') },
             ]}
             aside={
-                <div className="showcase-aside-block">
-                    {authIdentity && (
-                        <div className="showcase-aside-tabs">
-                            <div
-                                className={classNames(
-                                    'showcase-aside-tab',
-                                    'clickable',
-                                    !filterValues?.bookmarked && 'active',
-                                )}
-                                onClick={() => filterUpdate({ bookmarked: false })}
-                                onKeyDown={clickOnKeyEnter}
-                                tabIndex={0}
-                                aria-pressed={!filterValues.bookmarked}
-                                role="button">
-                                <em className="mdi mdi-tag-multiple-outline" />
-                                {translate('products.filters.all_products')}
-                            </div>
-                            <div
-                                className={classNames(
-                                    'showcase-aside-tab',
-                                    'clickable',
-                                    filterValues?.bookmarked && 'active',
-                                )}
-                                onClick={() => filterUpdate({ bookmarked: true })}
-                                onKeyDown={clickOnKeyEnter}
-                                role="button"
-                                tabIndex={0}
-                                aria-label={translate('products.filters.bookmarked')}
-                                aria-pressed={!!filterValues.bookmarked}>
-                                <em className="mdi mdi-cards-heart-outline" />
-
-                                {translate('products.filters.bookmarked')}
-                            </div>
-                        </div>
-                    )}
-
-                    <FormGroup
-                        id={'products_search'}
-                        label={translate('products.filters.search')}
-                        error={errors?.q}
-                        input={(id) => (
-                            <UIControlText
-                                value={filterValues.q}
-                                onChangeValue={(q: string) => filterUpdate({ q })}
-                                ariaLabel={translate('products.filters.search')}
-                                id={id}
-                                dataDusk="listProductsSearch"
-                            />
-                        )}
-                    />
-
-                    {labelsBlock}
-
-                    <ProductsFilterGroupProductCategories
-                        value={filterValues?.product_category_ids}
-                        setValue={(product_category_ids) => filterUpdate({ product_category_ids })}
-                        openByDefault={true}
-                        productCategories={productCategories}
-                        productCategoriesIconMap={productCategoriesIconMap}
-                    />
-
-                    <ProductsFilterGroupFunds
-                        value={filterValues?.fund_ids}
-                        funds={funds}
-                        setValue={(fund_ids) => filterUpdate({ fund_ids })}
-                        openByDefault={true}
-                        error={errors?.fund_ids}
-                    />
-
-                    <ProductsFilterGroupPrice
-                        filterValues={filterValues}
-                        filterUpdate={filterUpdate}
-                        errors={errors}
-                        toMax={toMax}
-                        openByDefault={true}
-                    />
-
-                    <ProductsFilterGroupProviders
-                        filterValues={filterValues}
-                        filterUpdate={filterUpdate}
-                        errors={errors}
-                    />
-
-                    <ProductsFilterGroupDistance values={filterValues} setValues={filterUpdate} errors={errors} />
-                    <ProductsFilterGroupReservationOptions value={filterValues} setValue={filterUpdate} />
-                    <ProductsFilterGroupPriceType value={filterValues} setValue={filterUpdate} />
-                </div>
+                <ProductsSidebarFilters
+                    errors={errors}
+                    filter={filter}
+                    filterValues={filterValues}
+                    filterUpdate={filterUpdate}
+                    funds={funds}
+                    productCategories={productCategories}
+                    productCategoriesIconMap={productCategoriesIconMap}
+                    providers={providers}
+                    showBookmarkTabs={!!authIdentity}
+                    toMax={toMax}
+                />
             }>
             {appConfigs && products && (
                 <Fragment>
